+++ packages/api/Tsz.Api/Persistence/AppDbContext.cs
@@ -1,6 +1,7 @@
 using Microsoft.EntityFrameworkCore;
 using Tsz.Modules.Contracts;
 using Tsz.Modules.Customers;
+using Tsz.Modules.LeaveTypes;
 using Tsz.Modules.Users;
 
 namespace Tsz.Api.Persistence;
@@ -16,5 +17,6 @@ public class AppDbContext : DbContext
         modelBuilder.ApplyConfigurationsFromAssembly(typeof(UsersModule).Assembly);
         modelBuilder.ApplyConfigurationsFromAssembly(typeof(CustomersModule).Assembly);
         modelBuilder.ApplyConfigurationsFromAssembly(typeof(ContractsModule).Assembly);
+        modelBuilder.ApplyConfigurationsFromAssembly(typeof(LeaveTypesModule).Assembly);
     }
 }